+++ src/dialog/enter_perennial_branches.go
@@ -8,21 +8,8 @@ import (
 	survey "gopkg.in/AlecAivazis/survey.v1"
 )
 
-// AskMainBranch lets the user enter the main branch.
-func AskMainBranch(mainBranch string, localBranches []string) (string, error) {
-	newMainBranch, err := AskForBranch(AskForBranchOptions{
-		Branches:      localBranches,
-		Prompt:        mainBranchPrompt(mainBranch),
-		DefaultBranch: mainBranch,
-	})
-	if err != nil {
-		return "", err
-	}
-	return newMainBranch, nil
-}
-
-// AskPerennialBranches lets the user enter the perennial branches.
-func AskPerennialBranches(localBranchesWithoutMain []string, perennialBranches []string) ([]string, error) {
+// EnterPerennialBranches lets the user enter the perennial branches.
+func EnterPerennialBranches(localBranchesWithoutMain []string, perennialBranches []string) ([]string, error) {
 	if len(localBranchesWithoutMain) == 0 {
 		return []string{}, nil
 	}
@@ -57,16 +44,6 @@ type askForBranchesOptions struct {
 	branches        []string
 	defaultBranches []string
 	prompt          string
-}
-
-func mainBranchPrompt(mainBranch string) string {
-	result := "Please specify the main development branch:"
-	currentMainBranch := mainBranch
-	if currentMainBranch != "" {
-		coloredBranch := color.New(color.Bold).Add(color.FgCyan).Sprintf(currentMainBranch)
-		result += fmt.Sprintf(" (current value: %s)", coloredBranch)
-	}
-	return result
 }
 
 func perennialBranchesPrompt(perennialBranches []string) string {
